bug fix to last bug fix :)

+++ rm_project.php
@@ -48,7 +48,7 @@ if (isset ($_REQUEST['project_id']) && getPriviledge () < 2)
 				
 				foreach (scandir ($path) as $file)
 				{
-				    if ($file != "." || $file != "..")
+				    if ($file != "." && $file != "..")
 				    {
 				        unlink ("$path/$file");
 				    }
